[PROJECT-WAR][1.0.0-SNAPSHOT][김대근] [BT] SP-110 #comment 회원등급관리 서비스 로직 변경 [RV] +review SR @spear79

+++ trunk/development-standard/standard-projects/project-war/src/main/java/standard/mvc/component/business/baroboard/user/manage/grade/service/UserGradeServiceImpl.java
@@ -1,10 +1,13 @@
 package standard.mvc.component.business.baroboard.user.manage.grade.service;
 
 import java.util.List;
+import java.util.Properties;
 
 import javax.annotation.Resource;
 
+import org.apache.commons.lang.StringUtils;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import standard.mvc.component.business.baroboard.user.manage.grade.vo.UserGradeManage;
 import standard.mvc.component.business.baroboard.user.manage.grade.vo.UserMenuByGrade;
@@ -38,6 +41,9 @@ public class UserGradeServiceImpl implements UserGradeService{
 	@Resource(name = "CoreService")
     private CoreService coreService;
 
+	@Resource(name = "fileUploadProperties")
+	private Properties fileUploadProperties;
+	
 	@Override
 	public List<UserGradeManage> inquiryUserGradeList(UserGradeManage userGradeManage) throws Exception {
 		if(null == userGradeManage || 0 == userGradeManage.getC_id()){
@@ -52,6 +58,13 @@ public class UserGradeServiceImpl implements UserGradeService{
 		
 		UserGradeManage rtnUserGradeManage = coreService.getNode(userGradeManage);
 		
+		if(StringUtils.isNotBlank(rtnUserGradeManage.getStoreFileNm())){
+			String filePath = fileUploadProperties.getProperty("grade.upload.dir");
+			filePath = filePath.replace("\\", "/");
+			
+			rtnUserGradeManage.setStoreFileNm(filePath + rtnUserGradeManage.getStoreFileNm());
+		}
+		
 		List<UserMenuByGrade> userMenuByGradeList = this.inquiryUserMenuByGradeList(rtnUserGradeManage.getC_id());
 		
 		rtnUserGradeManage.setUserMenuByGradeList(userMenuByGradeList);
@@ -59,19 +72,26 @@ public class UserGradeServiceImpl implements UserGradeService{
 		return rtnUserGradeManage;
 	}
 
+	@Transactional
 	@Override
 	public void saveUserGradeDetailInf(UserGradeManage userGradeManage) throws Exception {
 		UserGradeManage rtnUserGradeManage = coreService.addNode(userGradeManage);
-		int addGradeId = rtnUserGradeManage.getId();
+		int newGradeId = rtnUserGradeManage.getId();
 		
-		List<UserMenuByGrade> userMenuByGradeList = userGradeManage.getUserMenuByGradeList();
-		if(addGradeId > 0 && userMenuByGradeList != null){
-			for(UserMenuByGrade menu : userMenuByGradeList){
-				menu.setUserGradeId(addGradeId);
+		String[] menuList = userGradeManage.getAccessMenuList();
+		
+		if(menuList != null){
+			UserMenuByGrade userMenuByGrade = null;
+			for(String addMenu : menuList){
+				userMenuByGrade = new UserMenuByGrade();
 				
-				coreService.addNode(menu);
+				userMenuByGrade.setC_title(addMenu.substring(0, addMenu.indexOf("|")));
+				userMenuByGrade.setUserGradeId(newGradeId);
+				userMenuByGrade.setMenuId(Integer.parseInt(addMenu.substring(addMenu.indexOf("|") + 1)));
+				
+				coreService.addNode(userMenuByGrade);
 			}
-		}	
+		}
 	}
 
 	@Override
@@ -82,26 +102,63 @@ public class UserGradeServiceImpl implements UserGradeService{
 		return coreService.getChildNode(userMenuByGrade);
 	}
 
+	@Transactional
 	@Override
 	public void updateUserGradeInf(UserGradeManage userGradeManage) throws Exception {
 		coreService.alterNode(userGradeManage);
 		
-		List<UserMenuByGrade> userMenuByGradeList = userGradeManage.getUserMenuByGradeList();
-		if(userMenuByGradeList != null){
-			UserMenuByGrade rtnUserMenuByGrade = null;
+		UserMenuByGrade userMenuByGrade = new UserMenuByGrade();
+		userMenuByGrade.setUserGradeId(userGradeManage.getC_id());
+		List<UserMenuByGrade> userMenuByGradeList = coreService.getChildNode(userMenuByGrade);
+		
+		String[] userMenuByGradeIdArr = userGradeManage.getAccessMenuList();
+		if(userMenuByGradeList != null && userMenuByGradeIdArr != null){
+			UserMenuByGrade newMenu = null;
+			
+			//기존에 저장되어 있는 메뉴 리스트에 삭제된 메뉴를 찾아 삭제처리
 			for(UserMenuByGrade menu : userMenuByGradeList){
-				rtnUserMenuByGrade = coreService.getNode(menu);
+				boolean isExistsMenu = false;
 				
-				if(rtnUserMenuByGrade == null){
-					menu.setUserGradeId(userGradeManage.getC_id());
-					coreService.addNode(menu);
+				for(String addMenu : userMenuByGradeIdArr){
+					if(menu.getC_id() == Integer.parseInt(addMenu.substring(addMenu.indexOf("|") + 1))){
+						isExistsMenu = true;
+						break;
+					}
+				}
+				
+				if(!isExistsMenu){
+					coreService.removeNode(menu);
+				}
+			}
+			
+			//기존에 저장되어 있지 않는 메뉴인지 판별하여 신규 추가
+			for(String addMenu : userMenuByGradeIdArr){
+				boolean isExistsMenu = false;
+				for(UserMenuByGrade menu : userMenuByGradeList){
+					if(menu.getC_id() == Integer.parseInt(addMenu.substring(addMenu.indexOf("|") + 1))){
+						isExistsMenu = true;
+						break;
+					}
+				}
+				
+				if(!isExistsMenu){
+					newMenu = new UserMenuByGrade();
+					newMenu.setC_title(addMenu.substring(0, addMenu.indexOf("|")));
+					newMenu.setUserGradeId(userGradeManage.getC_id());
+					newMenu.setMenuId(Integer.parseInt(addMenu.substring(addMenu.indexOf("|") + 1)));
+					coreService.addNode(newMenu);
 				}
 			}
 		}
 	}
 
+	@Transactional
 	@Override
 	public void removeUserGradeInf(UserGradeManage userGradeManage) throws Exception {
-		coreService.removeNode(userGradeManage);
+		if(userGradeManage.getC_id() > 6){
+			coreService.removeNode(userGradeManage);
+		}else{
+			throw new RuntimeException("기본 제공 등급은 삭제할 수 없습니다.");
+		}		
 	}
 }
